+++ packages/curvenote-cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # curvenote
 
+## 0.17.4
+
+### Patch Changes
+
+- Updated dependencies [[`b2adb3e`](https://github.com/curvenote/curvenote/commit/b2adb3e78e5a2f4c860eeea0cb50bf5098d79bdb)]:
+  - @curvenote/common@0.9.0
+  - @curvenote/cli-plugin@0.17.4
+  - @curvenote/check-definitions@0.17.4
+  - @curvenote/check-implementations@0.17.4
+
 ## 0.17.3
 
 ### Patch Changes